refactor(home): align frontend module boundaries
