--- yak-ops-ui/src/pages/home/components/schedule-service.ts
@@ -1,1 +1,0 @@
-export * from '@/services/home';
