+++ yak-ops-ui/src/pages/home/types.ts
@@ -1,0 +1,12 @@
+import type { HomeDataCenterPeriod } from '@/services/home';
+
+export type HomeDataCenterTabKey = 'overview' | 'recent' | 'schedule';
+export type HomeDataCenterPeriodKey = HomeDataCenterPeriod;
+
+export interface HomeOverviewMetric {
+  label: string;
+  value: string;
+  compareLabel: string;
+  compareValue: string;
+  tone?: 'positive' | 'negative' | 'neutral';
+}
