+++ yak-ops-ui/src/services/data-service/index.ts
@@ -1,3 +1,4 @@
 export * from './api';
 export * from './constants';
+export * from './overview';
 export type * from './types';
